+++ src/models/user.ts
@@ -3,6 +3,7 @@ import mongoose, { Schema, Document } from "mongoose";
 export interface IUser extends Document {
   name: string;
   email: string;
+  role: "admin" | "user";
   emailVerified: Date | null;
   image: string | null;
   hashedPassword: string | null;
@@ -29,6 +30,7 @@ const userSchema = new Schema<IUser>(
   {
     name: { type: String, required: true },
     email: { type: String, required: true, unique: true },
+    role: { type: String, enum: ["admin", "user"], default: "user" },
     emailVerified: { type: Date, default: null },
     image: { type: String, default: null },
     hashedPassword: { type: String, default: null },