+++ src/types/index.ts
@@ -4,6 +4,7 @@ declare module "next-auth" {
   interface Session {
     user: {
       id: string;
+      role: "admin" | "user";
     } & DefaultSession["user"];
   }
 }